+++ contracts/Tokens/VAI/VAIController.sol
@@ -826,7 +826,7 @@ contract VAIController is VAIControllerInterface, VAIControllerStorageG4, VAICon
     }
 
     function getBlocksPerYear() public view returns (uint256) {
-        return 10512000; //(24 * 60 * 60 * 365) / 3;
+        return 21024000; //(24 * 60 * 60 * 365) / 1.5;
     }
 
     /**